Add hiring content for Core Engineers
I prefaced the list with 

> Weaviate Core Engineers typically have the following traits:

This is meant to indicate that the points on the list are not hard requirements, but "the more the better". For example, if someone would tick all boxes, but has never written a single line of Golang, this would not be an issue. Does that come across?